+++ src/api/profile.ts
@@ -45,17 +45,7 @@ export namespace ProfileApi {
       blocked: result.get('blockedArray') ?? [],
     };
 
-    const migrated: boolean | undefined = result.get('migrated');
     const kind: string | undefined = result.get('kind');
-
-    if (!migrated || !kind) {
-      console.log(`Detected legacy profile with ID '${result.id}'.`);
-      return {
-        kind: 'personal',
-        ...sharedProfileDetails,
-        ...constructCommonProfileDetails<PersonalProfileId>(result.id, result),
-      };
-    }
 
     if (kind === 'personal') {
       const profilePersonal: Parse.Object = result.get('profilePersonal');